requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiohttp-retry==2.9.1
aiosignal==1.4.0
aiosmtpd==1.4.6
asgiref==3.8.1
atpublic==6.0.1
attrs==25.3.0
azure-core==1.35.0
azure-identity==1.24.0
azure-storage-blob==12.26.0
boto3==1.20.26
botocore==1.23.54
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
decorator==4.4.2
dj-database-url==2.1.0
Django==5.2
django-anymail==10.2
django-ckeditor-5==0.2.18
django-cors-headers==3.14.0
django-jazzmin==2.6.0
django-phonenumber-field==8.1.0
django-sendgrid-v5==1.3.0
django-storages==1.12.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
ecdsa==0.19.1
environs==10.0.0
frozenlist==1.7.0
gunicorn==21.2.0
idna==3.6
imageio==2.37.0
imageio-ffmpeg==0.6.0
inflection==0.5.1
isodate==0.7.2
jmespath==0.10.0
MarkupSafe==3.0.2
marshmallow==3.20.1
moviepy==1.0.3
msal==1.33.0
msal-extensions==1.3.1
multidict==6.6.3
numpy==2.3.1
packaging==23.2
pdfrw==0.4
phonenumbers==9.0.11
pillow==11.3.0
proglog==0.1.12
propcache==0.3.2
psycopg2==2.9.9
psycopg2-binary==2.9.10
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-http-client==3.3.7
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.5.2
sendgrid==6.12.4
sentry-sdk==2.35.0
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.4
stripe==7.9.0
tqdm==4.67.1
twilio==9.7.0
typing_extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.18
Werkzeug==3.1.3
whitenoise==6.9.0
yarl==1.20.1
